Refactoring the build process to create a single library, libDelphi.a, along with a Python wrapper, DelphiPython
