+++ CMakeLists.txt
@@ -6,6 +6,9 @@ set(CMAKE_CXX_STANDARD_REQUIRED ON)
 set(CMAKE_MODULE_PATH ${CMAKE_MODULE_PATH} "${CMAKE_CURRENT_SOURCE_DIR}/cmake")
 set(Boost_NO_BOOST_CMAKE ON)
 set(CMAKE_POSITION_INDEPENDENT_CODE ON)
+set(CMAKE_CXX_FLAGS
+    "${CMAKE_CXX_FLAGS} -fvisibility=hidden -fvisibility-inlines-hidden")
+
 
 find_package(Graphviz)
 
@@ -28,29 +31,28 @@ set(PYBIND11_PYTHON_VERSION "${PYTHON_VERSION_MAJOR}.${PYTHON_VERSION_MINOR}")
 set(PYBIND11_CPP_STANDARD -std=c++17)
 add_subdirectory(external/pybind11)
 
-pybind11_add_module(kde lib/kde_wrapper.cpp)
-pybind11_add_module(AnalysisGraph lib/AnalysisGraph_wrapper.cpp)
-
-add_library(AnalysisGraphCpp lib/AnalysisGraph.cpp)
-target_link_libraries(AnalysisGraphCpp
-                      PRIVATE ${Boost_LIBRARIES}
-                              fmt::fmt
-                              "sqlite3"
-                              ${GRAPHVIZ_LIBS})
-
-target_link_libraries(kde PRIVATE ${Boost_LIBRARIES} fmt::fmt)
-target_link_libraries(AnalysisGraph
-                      PRIVATE ${Boost_LIBRARIES} fmt::fmt ${GRAPHVIZ_LIBS})
+file(GLOB SRC_FILES ${PROJECT_SOURCE_DIR}/lib/*.cpp)
 
 message("-- Using Eigen3 version 3.3 bundled with Delphi")
 include_directories(external/eigen)
 
 find_package(SQLite3 REQUIRED)
 include_directories(${SQLITE3_INCLUDE_DIRS})
-target_link_libraries(AnalysisGraph PRIVATE "sqlite3")
+
+add_library(Delphi ${SRC_FILES})
+target_link_libraries(Delphi
+                      PRIVATE ${Boost_LIBRARIES}
+                              fmt::fmt
+                              "sqlite3"
+                              ${GRAPHVIZ_LIBS}
+                              )
+
 
 # Tests
 add_library(doctest INTERFACE)
 add_executable(cpptests tests/cpptests.cpp)
 target_include_directories(doctest INTERFACE external/doctest)
-target_link_libraries(cpptests PRIVATE doctest AnalysisGraphCpp)
+target_link_libraries(cpptests PRIVATE doctest Delphi)
+
+pybind11_add_module(DelphiPython MODULE lib/wrappers/AnalysisGraph_wrapper.cpp)
+target_link_libraries(DelphiPython PRIVATE Delphi)
