+++ lib/kde.cpp
@@ -1,11 +1,11 @@
 #include <random>
 #include <boost/range/irange.hpp>
 #include <boost/random/mersenne_twister.hpp>
+#include <boost/range/numeric.hpp>
+#include <boost/range/adaptors.hpp>
 #include <boost/range/algorithm/for_each.hpp>
 #include <boost/lambda/bind.hpp>
 #include <boost/lambda/lambda.hpp>
-#include "utils.hpp"
-#include "rng.hpp"
 #include "kde.hpp"
 
 using std::vector;
@@ -24,6 +24,21 @@ double sample_from_normal(
   normal_distribution<> d{mu, sd};
   return d(gen);
 }
+
+KDE::KDE(std::vector<double> v) : dataset(v) {
+    using boost::adaptors::transformed;
+    using boost::lambda::_1;
+    using utils::mean;
+
+    // Compute the bandwidth using Silverman's rule
+    mu = mean(v);
+    auto X = v | transformed(_1 - mu);
+
+    // Compute standard deviation of the sample.
+    size_t N = v.size();
+    double stdev = sqrt(inner_product(X, X, 0.0) / (N - 1));
+    bw = pow(4 * pow(stdev, 5) / (3 * N), 1 / 5);
+  }
 
 vector<double> KDE::resample(int n_samples) {
   vector<double> samples;
@@ -55,3 +70,4 @@ vector<double> KDE::pdf(vector<double> v) {
 }
 
 double KDE::logpdf(double x) { return log(pdf(x)); }
+
